Add CautiousSeeker
CautiousSeeker works similar to Opportunist, but instead of running away
it seeks enemy pieces. To avoid being eaten it never goes too close and only
eats enemy pieces if they approach it on their own.

This means that it will still draw against runner, opportunist and itself, but
increases the chance of eating "dumber" player's pieces compared to Opportunist.